work in progress: dynamicly match route to scenes by user
